requirements.txt
alabaster==0.7.12
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attr==0.3.1
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
black==22.1.0
bleach==4.1.0
boltons==21.0.0
Brotli==1.0.9
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
configparser==5.2.0
cryptography==36.0.1
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
Django==3.2.5
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-jazzmin==2.4.9
djangorestframework==3.13.1
dl==0.1.0
docopt==0.6.2
docutils==0.17
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
funcsigs==1.0.2
HTMLParser==0.0.2
ghostscript==0.7
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.1
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jnius==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.5.0
lockfile==0.12.2
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
mock==4.0.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.1
opencv-python==4.6.0.66
openpyxl==3.0.9
ordereddict==1.1
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pdfminer.six==20220524
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyPDF2==2.1.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyunpack==0.3
pyzmq==22.3.0
railroad==0.5.0
requests==2.27.1
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.2.0
tabula-py==2.4.0
tabulate==0.8.9
terminado==0.13.1
testpath==0.5.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
unicodedata2==14.0.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
yarg==0.1.9
zipp==3.7.0
